Refactor variables to float for trapezoid area calculation

+++ Area do trapezio.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(int argc, char** argv)
 {
 
-	int baseMenor = 0, altura = 0, baseMaior = 0;
+	float baseMenor = 0, altura = 0, baseMaior = 0;
 	cout << "##### AREA DO TRAPEZIO #####\n\n >>> ((base Maior + base menor) x altura)/2 <<<\n\n";
 	
 	cout << "Informe o valor da base menor (em metros)? -> ";
@@ -14,7 +14,7 @@ int main(int argc, char** argv)
 	cout << "Informe o valor da altura (em metros)? ->";
 	cin >> altura;
 
-	/*Exibindo o valor da área calculada*/	
+	/*Exibindo o valor da Ã¡rea calculada*/	
 	cout << "\n Area do Trapezio -> (("<<baseMenor<<"m + "<<baseMaior<<"m)x "<<altura<<"m)/2 = "<< (((baseMaior + baseMenor) * altura)/2)<< "m2";
 	return 0;
 }
